requirements.txt
flask==3.1.3
flask-cors==6.0.2
requests==2.32.5
pandas==3.0.1
gunicorn==21.2.0
urllib3==2.6.3
openpyxl>=3.1.0
google-generativeai>=0.8.0
